+++ src/phrases/phrases-nb.js
@@ -584,7 +584,7 @@ const phrases = {
       title: 'Status',
       change: 'Endre status',
       qualityAssurance: 'Kvalitetssikring',
-      saveAsNew: 'Lagre kopi som ny artikkel',
+      saveAsNew: 'Lag kopi',
       deleteLanguageVersion: {
         button: 'Slett {languageVersion} versjon',
         title: 'Slett språkversjon',
