+++ src/phrases/phrases-nn.js
@@ -593,7 +593,7 @@ const phrases = {
       title: 'Status',
       change: 'Endre status',
       qualityAssurance: 'Kvalitetssikring',
-      saveAsNew: 'Lagre kopi som ny artikkel',
+      saveAsNew: 'Lag kopi',
       deleteLanguageVersion: {
         button: 'Slett språkversjon',
         title: 'Slett språkversjon',
